fix compression enabling logic #31
The compression-related code in dump/tape.c is under:
	#ifdef HAVE_BLOCK_TRANSFORMATION
but nothing sets that.  Dump thus claims to compress, the file header
is set up for compression but nothing is in fact compressed; restore
subsequently fails to restore as it can't decompress.

Adjust transformation.h to define HAVE_BLOCK_TRANSFORMATION when one
of the relevant HAVE_xyz defines are found.

Fixes: https://sourceforge.net/p/dump/patches/31